P2.0.7_a4"
I hated stripping out the [delay: xxx] from the SmartThings platform as this is used to help prevent cut-offs, etc with back to back notifications.  I've added it back into the def playTrackAndRexxxe() functions but with a detection for SmartThings/Hubitat where Hubitat will not use it (since it doesn't support it apparently).